[TASK] modify controller actions. put search in dealers. use page renderer to get JS labels.
