+++ Classes/Controller/DealersController.php
@@ -26,7 +26,11 @@ namespace Pixelant\PxaDealers\Controller;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
+use Pixelant\PxaDealers\Domain\Model\Dealer;
+use Pixelant\PxaDealers\Domain\Model\Demand;
 use Pixelant\PxaDealers\Domain\Model\Search;
+use TYPO3\CMS\Core\Page\PageRenderer;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  *
@@ -35,8 +39,19 @@ use Pixelant\PxaDealers\Domain\Model\Search;
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  *
  */
-class DealersController extends AbstractConroller
+class DealersController extends AbstractController
 {
+    /**
+     * Map action initialize
+     */
+    public function initializeMapAction()
+    {
+        $pageRenderer = GeneralUtility::makeInstance(PageRenderer::class);
+        $pageRenderer->addInlineLanguageLabelFile(
+            'EXT:pxa_dealers/Resources/Private/Language/locallang.xlf',
+            'js.'
+        );
+    }
 
     /**
      * action map
@@ -46,6 +61,58 @@ class DealersController extends AbstractConroller
      */
     public function mapAction(Search $search = null)
     {
-        $this->renderMap($search);
+        $demand = Demand::getInstance($this->settings['demand']);
+
+        if ($search !== null) {
+            $search->setSearchFields(GeneralUtility::trimExplode(
+                ',',
+                $this->settings['search']['searchFields'],
+                true
+            ));
+            $demand->setSeach($search);
+
+            if ($search->isSearchInRadius() && !empty($this->settings['map']['googleServerApiKey'])) {
+                if (empty($search->getLat()) || empty($search->getLng())) {
+                    // Get from address
+                    list($lat, $lng) = $this->getAddressInfo($search->getSearchTermOriginal());
+                } else {
+                    // Use user position
+                    $lat = $search->getLat();
+                    $lng = $search->getLng();
+                }
+
+                if ($lat && $lng) {
+                    $search->setLat($lat);
+                    $search->setLng($lng);
+                    $search->setRadius(intval($this->settings['search']['radius']) ?: 100);
+
+                    $searchCenter = [
+                        'lat' => $lat,
+                        'lng' => $lng
+                    ];
+                } else {
+                    $search->setSearchInRadius(false);
+                }
+            }
+        }
+
+        $allCategoriesUids = [];
+        $allCountriesUids = [];
+
+        $dealers = [];
+        /** @var Dealer $dealer */
+        foreach ($this->dealerRepository->findDemanded($demand) as $dealer) {
+            $dealers[$dealer->getUid()] = $dealer;
+
+            $allCategoriesUids = array_merge($allCategoriesUids, $dealer->getCategoriesAsUidsArray());
+            $allCountriesUids[] = $dealer->getCountryUid();
+        }
+
+        $this->view->assignMultiple([
+            'dealers' => $dealers,
+            'allCategoriesUids' => implode(',', array_unique($allCategoriesUids)),
+            'allCountriesUids' => implode(',', array_unique($allCountriesUids)),
+            'searchCenter' => isset($searchCenter) ? $searchCenter : ['lat' => 0, 'lng' => 0]
+        ]);
     }
 }
