+++ Resources/Private/Language/locallang_db.xlf
@@ -81,9 +81,6 @@
             <trans-unit id="flexform.actions.filters">
                 <source>Filters</source>
             </trans-unit>
-            <trans-unit id="flexform.actions.searchResults">
-                <source>Search results</source>
-            </trans-unit>
             <trans-unit id="flexform.actions.searchSection">
                 <source>Search</source>
             </trans-unit>
