+++ Resources/Private/Partials/Dealers/List.html
@@ -1,7 +1,7 @@
 <html xmlns:f="http://typo3.org/ns/TYPO3/CMS/Fluid/ViewHelpers" xmlns:pxa="http://typo3.org/ns/Pixelant/PxaDealers/ViewHelpers" data-namespace-typo3-fluid="true">
 <f:for each="{dealers}" as="dealer">
     <div id="dealer-{dealer.uid}"
-         class="dealer-item isotope-item{f:format.raw(value: ' cat-{category.uid}') -> f:for(each: dealer.categories, as: 'category')} country-{dealer.country}"
+         class="dealer-item isotope-item{f:format.raw(value: ' cat-{category.uid}') -> f:for(each: dealer.categories, as: 'category')} country-{dealer.country.uid}"
          data-uid="{dealer.uid}"
          data-marker-type="default">
