+++ Resources/Private/Partials/Javascript/JsMapSettings.html
@@ -11,7 +11,6 @@
         };
 
         PxaDealersMaps.settings = <f:format.htmlentitiesDecode><pxa:jsonEncode value="{settings.map}" /></f:format.htmlentitiesDecode>;
-	    PxaDealersMaps.labels = <f:format.htmlentitiesDecode><pxa:jsonEncode value="{labelsJs}" /></f:format.htmlentitiesDecode>;
 
         w.PxaDealersMaps = PxaDealersMaps;
     })(window);
