--- Resources/Private/Templates/Search/SearchResults.html
@@ -1,9 +1,0 @@
-<div xmlns="http://www.w3.org/1999/xhtml" lang="en"
-     xmlns:f="http://typo3.org/ns/TYPO3/Fluid/ViewHelpers">
-
-    <f:layout name="Default" />
-
-    <f:section name="main">
-        <f:render partial="MapTemplate" arguments="{_all}"/>
-    </f:section>
-</div>
