--- Resources/Private/Templates/Search/Suggest.html
@@ -1,6 +1,0 @@
-<html xmlns:f="http://typo3.org/ns/TYPO3/CMS/Fluid/ViewHelpers" xmlns:pxa="http://typo3.org/ns/Pixelant/PxaDealers/ViewHelpers" data-namespace-typo3-fluid="true">
-
-    <f:layout name="Empty" />
-
-    <f:section name="main"><pxa:jsonEncode>{data}</pxa:jsonEncode></f:section>
-</html>
